added day2-3

+++ Day3/Day3-1.py
@@ -1,4 +1,4 @@
-users_temperature = int(input("What is your temperature? "))
+users_temperature = float(input("What is your temperature? "))
 if users_temperature < 35:
     print("not too cold")
 elif users_temperature > 37:
